BACKEND/hems_backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-core>=0.1.0

dj-database-url>=2.1.0
django-filter>=24.0
reportlab>=4.0.0
